requirements.txt
streamlit
yt-dlp
webvtt-py
google-generativeai
fpdf
langchain-google-genai
